requirements.txt
google-cloud-storage
line-bot-sdk
googletrans==3.1.0a0
uvicorn[standard]
fastapi